% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTexpSimulation.R
\name{CTexpSimulation}
\alias{CTexpSimulation}
\title{Simulate cell type specific expression analysis}
\usage{
CTexpSimulation(
  g = 1000,
  n = 200,
  p = 0.05,
  m_lfc = 4,
  sd_lfc = 0.2,
  parameters = simulation.parameters
)
}
\arguments{
\item{g}{number of genes - defaults to 1,000}

\item{n}{number of samples - defaults to 200, an event spliut between cases and controls}

\item{p}{proportion of genes that have non-zero LFC in each cell type - defaults to 0.05}

\item{m_lfc}{mean of the normal distribution from which the absolute lfc values are drawn - dafaults to 4}

\item{sd_lfc}{sd of the normal distribution from which the absolute lfc values are drawn - dafaults to 0.2}

\item{parameters}{additional parameters used to simulate expression. See the simulation.parameters document for more details.}
}
\value{
A list with the following objects:
\enumerate{
\item bulk.expression - a g X n matrix of the expression values in mixted samples
\item ols.mixture.estimates - a n X k matrix of cell type proportions estimated using ordinary least squares
\item true.mixture - a n X k matrix of true cell type proportions.
\item ct.reference - a g X k matrix of reference values for cell type specific expression
\item simulated.lfc - a g X k matrix of simulated lfc in each cell type.
}
}
\description{
This function is based on the simulations described in a benchmarking paper by Meng et al., 2023 (https://doi.org/10.1093/bib/bbac516)
}
