% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTexpSimulation.R
\docType{data}
\name{data-simulation.parameters}
\alias{data-simulation.parameters}
\alias{simulation.parameters}
\title{#' This is data to be included in my package}
\format{
An object of class \code{list} of length 6.
}
\usage{
simulation.parameters
}
\description{
A list of default parameters for cell type-specific expression simulation. There are g genes, k cell types, and n samples evenly divided between cases and controls.
mean_hat is a vector length k, with mean expression parameters for each cell-types, disp_hat is also a vector of length k with the dispersion parameters for each cell type.
mean_covar is a k x k matrix giving the covariance between the mean parameters in each cell type, disp_hat gives the covariance between the dispersion parameters in the different cell types.
alpha_cases and alpha_controls are k length vectors with parameters for the dirichlet distribution from which cell type proportion are drawn.
These parameters were obtained from the Supplementary materials of Meng et al., 2023 (https://doi.org/10.1093/bib/bbac516)
}
\keyword{data}
