% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrema.R
\docType{data}
\name{data-test.set}
\alias{data-test.set}
\alias{test.set}
\title{#' This is data to be included in my package}
\format{
An object of class \code{list} of length 2.
}
\usage{
test.set
}
\description{
A list containing a dataframe of genes, effect sizes and effect size variances, and a list with a single gene set.
Access using test.set$postdata and test.set$gs respectively. This data is used to carry out tests of the mrema function.
}
\keyword{data}
