% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrema.R
\name{mrema}
\alias{mrema}
\title{Run gene set analysis using Mixture of Random-Effects Meta-Analysis}
\usage{
mrema(
  postdata,
  raw.gs,
  DF = NULL,
  threshold = NULL,
  ncores = 1,
  overlap = 0.25
)
}
\arguments{
\item{postdata}{A three column dataframe with columns corresponding to gene name, lfc and lfc variance, respectively}

\item{raw.gs}{A list of named gene sets.}

\item{DF}{Degrees of freedom between the null and alternative hypotheses.}

\item{threshold}{The foldchange above which a gene is considered differentially expressed.}

\item{ncores}{The number of cores to use for parallel computing. Defaults to 1.}

\item{overlap}{The proportion of the DE normal distributions. Defaults to 0.25}
}
\value{
A list of two dataframes, results containing the results of the tests and parameters containing the parameters of the distribution of the gene set log fold changes.
}
\description{
Run gene set analysis using Mixture of Random-Effects Meta-Analysis
}
