% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REtest.R
\name{REtest}
\alias{REtest}
\title{Run gene set analysis using Random-Effect permutations}
\usage{
REtest(estimates, lfc_thresh, gene_set, type = "DE")
}
\arguments{
\item{estimates}{A three column dataframe with columns corresponding to gene name, lfc and lfcSE, respectively}

\item{lfc_thresh}{The log2-foldchange above which a gene is considered differentially expressed.}

\item{gene_set}{A list of named gene sets.}

\item{type}{Which test to conduct, for differential expression, upregulation, downregulation, "DE", "upreg" and "downreg" respectively}
}
\value{
A data.frame with gene set name, set size, proportion of distribution above the threshold and p-value.
}
\description{
Run gene set analysis using Random-Effect permutations
}
